refactor: apply new sidebar to all sidebar occurances
