--- src/shared/components/new-sidebar.tsx
@@ -1,48 +1,0 @@
-import {
-  Sidebar,
-  SidebarContent,
-  SidebarGroup,
-  SidebarGroupContent,
-  SidebarGroupLabel,
-  SidebarMenu,
-  SidebarMenuButton,
-  SidebarMenuItem
-} from '@/components/ui/sidebar'
-
-export type SidebarItem = {
-  label: string
-  icon: React.ReactNode
-  key: string
-  onClick?: () => void
-}
-
-type SidebarProps = {
-  items: SidebarItem[]
-}
-
-export function NewSidebar({ items }: SidebarProps) {
-  return (
-    <Sidebar variant="floating" collapsible="icon">
-      <SidebarContent>
-        <SidebarGroup>
-          <SidebarGroupLabel>Documentação</SidebarGroupLabel>
-          <SidebarGroupContent>
-            <SidebarMenu>
-              {items.map((item) => (
-                <SidebarMenuItem key={item.label}>
-                  <SidebarMenuButton asChild>
-                    <a href={item.key}>
-                      {item.icon}
-                      <span>{item.label}</span>
-                    </a>
-                  </SidebarMenuButton>
-                </SidebarMenuItem>
-              ))}
-            </SidebarMenu>
-          </SidebarGroupContent>
-        </SidebarGroup>
-      </SidebarContent>
-    </Sidebar>
-    
-  )
-}
